+++ src/navwebserver/SDDBT.h
@@ -4,9 +4,9 @@
 #include <Arduino.h>
 
 struct SDDBT_Data {
-  float water_depthFeet;
-  float water_depthMeters;
-  float water_depthFathoms;
+  int water_depthFeet;
+  int water_depthMeters;
+  int water_depthFathoms;
 };
 
 SDDBT_Data parseSDDBT(String nmea);
