+++ src/navwebserver/WIMWV.h
@@ -4,9 +4,9 @@
 #include <Arduino.h>
 
 struct WIMWV_Data {
-  float   wind_angle;
+  int   wind_angle;
   char    wind_angleReference;     // R/T
-  float   wind_speedKts;
+  int   wind_speedKts;
   char    wind_speedUnit; // N/M/K
   char    wind_sensorStatus;        // A/V
 };
